requirements.txt
streamlit==1.28.1
groq==0.11.0
python-dotenv==1.0.0
httpx==0.27.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
duckduckgo-search==6.3.5
